+++ spec/ios_download_strings_files_from_glotpress_spec.rb
@@ -7,7 +7,7 @@ describe Fastlane::Actions::IosDownloadStringsFilesFromGlotpressAction do
   let(:locales_subset) { { 'fr-FR': 'fr', 'zh-cn': 'zh-Hans' } }
 
   def gp_stub(locale:, query:)
-    stub_request(:get, "#{gp_fake_url}/#{locale}/default/export-translations").with(query: query)
+    stub_request(:get, "#{gp_fake_url}/#{locale}/default/export-translations/").with(query: query)
   end
 
   describe 'downloading export files from GlotPress' do
@@ -81,7 +81,7 @@ describe Fastlane::Actions::IosDownloadStringsFilesFromGlotpressAction do
         # Assert
         expect(stub).to have_been_made.once
         expect(File).not_to exist(File.join(tmp_dir, 'Base.lproj', 'Localizable.strings'))
-        expect(error_messages).to eq(['Error downloading locale `unknown-locale` — 404 Not Found'])
+        expect(error_messages).to eq(["Error downloading locale `unknown-locale` — 404 Not Found (#{gp_fake_url}/unknown-locale/default/export-translations/?filters%5Bstatus%5D=current&format=strings)"])
       end
     end