+++ spec/shared_examples_for_update_metadata_source_action.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 RSpec.shared_examples 'update_metadata_source_action' do |options|
   it 'updates any block in a given .po file with the values from the given sources' do
-    Dir.mktmpdir do |dir|
+    in_tmp_dir do |dir|
       output_path = File.join(dir, 'output.po')
       dummy_text = <<~PO
         msgctxt "key1"
@@ -45,7 +45,7 @@ RSpec.shared_examples 'update_metadata_source_action' do |options|
   it 'combines the given `release_version` and `whats_new` parameter into a new block' do
     pending 'this currently fails; in the long run, we might consolidate `whats_new` with `release_notes`' if options[:whats_new_fails]
 
-    Dir.mktmpdir do |dir|
+    in_tmp_dir do |dir|
       output_path = File.join(dir, 'output.po')
       dummy_text = <<~PO
         msgctxt "v1.0-whats-new"
@@ -80,7 +80,7 @@ RSpec.shared_examples 'update_metadata_source_action' do |options|
   it 'adds entries passed as input even if not part of the original `.po` file' do
     pending 'this currently fails and will be addressed as part of the upcoming refactor/rewrite of the functionality'
 
-    Dir.mktmpdir do |dir|
+    in_tmp_dir do |dir|
       output_path = File.join(dir, 'output.po')
       dummy_text = <<~PO
         msgctxt "key1"
@@ -119,7 +119,7 @@ RSpec.shared_examples 'update_metadata_source_action' do |options|
   end
 
   it 'combines the given `release_version` and `release_notes` in a new block, keeps the n-1 ones, and deletes the others' do
-    Dir.mktmpdir do |dir|
+    in_tmp_dir do |dir|
       output_path = File.join(dir, 'output.po')
       dummy_text = <<~PO
         msgctxt "release_note_0122"